requirements.txt
python-dotenv
spacy
pgvector
psycopg2
